added training_pipeline.py and main_train.py and updated README.md

+++ 02.Student_Performance_ML_E2E_project/src/student_performace_MLProject/pipelines/prediction_pipeline.py
@@ -41,8 +41,8 @@ class CustomData:
          parental_level_of_education:str,
          lunch:str,
          test_preparation_course:str,
-         reading_score:int,
-         writing_score:int):
+         reading_score:float,
+         writing_score:float):
           
 
         self.gender = gender
